requirements.txt
streamlit==0.80.0
annoy==1.17.0
fuzzywuzzy==0.18.0
tqdm==4.60.0
tensorflow-cpu==2.4.1
pandas==1.2.4
matplotlib==3.4.1
requests==2.25.1
numpy==1.19.5
beautifulsoup4==4.9.3
Pillow==8.2.0
